requirements.txt
asgiref==3.7.2
Django==5.0.1
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
mysqlclient==2.2.1
pillow==10.2.0
pytz==2023.4
shell==1.0.1
sqlparse==0.4.4
tzdata==2023.4
whitenoise==6.6.0